add arena and bitset (#63)
* replace slotmap with a custom arena implementation and implement bitsets

the old slotmap is dead, long live the new slotmap!

`slotmap` was implemented with the expectation that keys would be
removed from the maps. this isn't the case for us, which means that a
key contains a bunch of extra data we don't care about. with this new
implementation, we can convert any key directly into a `usize` index,
allowing us to do things with data structures such as BDDs and bitsets
much more easily.

* fix display implementation for ArgumentsStrategy